Added dcf77 library for 8-bit Pinguino and Timer4 support for Pinguino 26j50 Fixed lcd, i2c and interrupt library for 8-bit Pinguino

+++ branches/x.3/examples/10.Libraries/EEPROM/eeprom_clear.pde
@@ -9,7 +9,7 @@
 
 void setup()
 {
-	int i;
+	u16 i;
 	// write a 0 to all 256 bytes of the EEPROM
 	for (i = 0; i < 256; i++)
 		EEPROM.write(i, 0);
